worked on improved out_of_order detection
the OutOfOrder Network now detects which object is out of order and needs to be moved. Currently this is simplified to only one object to be out of order (untested with more than one). This also detects which object is more optimal to move, meaning that it will move the object with more space in between the remaining objects to make putting down easier for the robot. Tested with simulated input and bot. One issue: sometimes the script needs to be started again because all values of the odd-ensemble stay zero, which leads to no evidence being accumulated